pdf-extractor-streamlit/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
Pillow
PyMuPDF
pytesseract
opencv-python-headless
pandas
openpyxl
python-multipart
aiofiles
